refactoring: replaced registrationform to login page, added styles for registrationform
