+++ src/components/AppRouter.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Route, Router, Routes } from "react-router-dom";
 import Login from "./Login";
-import Registration from './Registration';
 import MapPage from './MapPage';
 import Profile from "./Profile";
 import { RequireAuth } from "./RequireAuth"
@@ -10,7 +9,8 @@ const AppRouter = () => {
 
     return (
         <Routes>
-            <Route path="/login" element={<Login />} />
+            <Route path="/login" element={<Login formType="login"/>} />
+            <Route path="/registration" element={<Login formType="registration"/>} />
             <Route path="/map" element={
                 <RequireAuth>
                     <MapPage />
@@ -21,8 +21,7 @@ const AppRouter = () => {
                     <Profile />
                 </RequireAuth>
             } />
-            <Route path="/registration" element={<Registration />} />
-            <Route path="*" element={<Login />} />
+            <Route path="*" element={<Login formType="login"/>} />
         </Routes>
     )
 }
